CLINT: Added tests Added Code Coverage Added some more docs
