+++ clint_smith/.env.sample
@@ -1,0 +1,14 @@
+# .env
+
+# Optional
+# Makefile command for SQL Server Run
+SQL_CONTAINER=sql2022
+SQL_PASSWORD=YOUR_PASSWORD
+SQL_IMAGE=mcr.microsoft.com/mssql/server:2022-latest
+SQL_PORT=1433
+# End Optional
+
+# These are necessary
+EDI_TOKEN=YOUR_TOKEN
+SQL_CONN_STRING=Server=localhost,1433;Database=HIPAA_5010_837P;User Id=sa;Password=YOUR_PASSWORD;TrustServerCertificate=True;
+STAGING_CONN_STRING=Server=localhost,1433;Database=CLAIM_STAGING;User Id=sa;Password=YOUR_PASSWORD;TrustServerCertificate=True;
