+++ clint_smith/docs/env-setup.md
@@ -1,0 +1,3 @@
+# .env setup
+1. Copy [*.env.sample*](../.env.sample) and rename to .env
+2. Replace passwords and token with your specifics
