+++ clint_smith/tests/EDI837.Ingestion.Tests/EDI837.Ingestion.Tests.csproj
