--- clint_smith/tests/EDI837.Ingestion.Tests/UnitTest1.cs
@@ -1,10 +1,0 @@
-namespace EDI837.Ingestion.Tests;
-
-public class UnitTest1
-{
-    [Fact]
-    public void Test1()
-    {
-
-    }
-}
