+++ clint_smith/tests/EDI837.Ingestion.Tests/samples/837-sample-file-fail-no-iea.edi
@@ -1,0 +1,44 @@
+ISA*00*          *00*          *ZZ*1234567        *ZZ*11111          *170508*1141*^*00501*000000101*1*P*:~
+GS*HC*XXXXXXX*XXXXX*20170617*1741*101*X*005010X222A1~
+ST*837*1239*005010X222A1~
+BHT*0019*00*010*20170617*1741*CH~
+NM1*41*2*SUBMITTER*****46*ABC123~
+PER*IC*BOB SMITH*TE*4805551212~
+NM1*40*2*RECEIVER*****46*44556~
+HL*1**20*1~
+NM1*85*2*BILLING PROVIDER*****XX*1122334455~
+N3*1234 SOME ROAD~
+N4*CHICAGO*IL*606739999~
+REF*EI*999999999~
+HL*2*1*22*0~
+SBR*P*18*******12~
+NM1*IL*1*BLOGGS*JOE****MI*1234567890~
+N3*1 SOME BLVD~
+N4*CHICAGO*IL*606129998~
+DMG*D8*19570111*M~
+NM1*PR*2*PAYER*****PI*12345~
+N3*1 PAYER WAY~
+N4*ST LOUIS*MO*212441850~
+REF*2U*W1014~
+CLM*1000A*140***19:B:1*Y*A*Y*Y~
+HI*ABK:I10~
+LX*1~
+SV1*HC:99213*140*UN*1***1~
+DTP*472*D8*20151124~
+HL*3*1*22*0~
+SBR*P*18*******12~
+NM1*IL*1*BLOGGS*FRED****MI*9876543201~
+N3*1 ANOTHER STR~
+N4*CHICAGO*IL*606129998~
+DMG*D8*19700601*M~
+NM1*PR*2*PAYER*****PI*12345~
+N3*1 PAYER WAY~
+N4*ST LOUIS*MO*212441850~
+REF*2U*W1014~
+CLM*1001A*140***19:B:1*Y*A*Y*Y~
+HI*ABK:I10~
+LX*1~
+SV1*HC:99213*140*UN*1***1~
+DTP*472*D8*20151124~
+SE*41*1239~
+GE*1*101~
